requirements.txt
streamlit>0.66
torch
torchvision
scikit-image
pillow
glob
numpy